third_party/blink/web_tests/platform/generic/fast/alignment/overwrite-content-alignment-expected.txt

Test overwriting justify-content and align-content works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(childOfDefaultParent, '').getPropertyValue('justify-content') is "unsafe flex-end"
PASS getComputedStyle(childOfDefaultParent, '').getPropertyValue('align-content') is "safe center"
PASS getComputedStyle(childOfParentId, '').getPropertyValue('justify-content') is "space-between"
PASS getComputedStyle(childOfParentId, '').getPropertyValue('align-content') is "space-around"
PASS successfullyParsed is true

TEST COMPLETE
